requirements.txt
pandas
pyarrow
minio
psycopg2-binary
jupyter